cleanup nu-command, remove redundant code
