+++ crates/nu-command/src/filesystem/mod.rs
@@ -16,7 +16,8 @@ pub use glob::Glob;
 pub use ls::Ls;
 pub use mkdir::Mkdir;
 pub use mv::Mv;
-pub use open::{BufferedReader, Open};
+pub use open::Open;
 pub use rm::Rm;
 pub use save::Save;
 pub use touch::Touch;
+pub use util::BufferedReader;
