+++ crates/nu-command/src/network/fetch.rs
@@ -1,3 +1,5 @@
+use crate::BufferedReader;
+
 use base64::encode;
 use nu_engine::CallExt;
 use nu_protocol::ast::Call;
@@ -10,7 +12,7 @@ use nu_protocol::{
 use reqwest::blocking::Response;
 
 use std::collections::HashMap;
-use std::io::{BufRead, BufReader, Read};
+use std::io::BufReader;
 
 use reqwest::StatusCode;
 use std::path::PathBuf;
@@ -335,34 +337,6 @@ fn helper(
             ),
             span,
         )),
-    }
-}
-
-pub struct BufferedReader<R: Read> {
-    input: BufReader<R>,
-}
-
-impl<R: Read> Iterator for BufferedReader<R> {
-    type Item = Result<Vec<u8>, ShellError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let buffer = self.input.fill_buf();
-        match buffer {
-            Ok(s) => {
-                let result = s.to_vec();
-
-                let buffer_len = s.len();
-
-                if buffer_len == 0 {
-                    None
-                } else {
-                    self.input.consume(buffer_len);
-
-                    Some(Ok(result))
-                }
-            }
-            Err(e) => Some(Err(ShellError::IOError(e.to_string()))),
-        }
     }
 }
